backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
sqlalchemy==2.0.23
